Alterações no component presentation e na aba inicio
